fix: resolve 15 code-review findings across fusion-mlx
Code-review fixes (#63-#78):
- api/utils.py: extract_multimodal_content preserves video/video_url/
  audio_url parts; has_multimodal detects them (#70)
- api/openai_models.py: ContentPart gains video/audio_url fields +
  AudioURL model, parity with api/models.py (#77)
- api/models.py: CompletionRequest.logprobs bool->int (OpenAI legacy
  spec); declare n/best_of/echo/response_format/stream_options
  (previously silently dropped -> AttributeError cascade)
- service/helpers.py: _extract_token_logprob top_k<=0 early-return
  guard (#72)
- pool/memory_enforcer.py: split _is_emergency_pressure (side-effect)
  from pure _evaluate_emergency_pressure; recheck uses pure variant
  to avoid double-incrementing sustained-pressure counter (#73)
- pool/engine_pool.py: detach/settle split - unload detaches under
  lock + settles outside; sync unload_engine subtracts estimated_size
  (#69, #74)
- cli_serve.py: _serve_from_model_dir honors --host/--port (#75)
- video/ltx2/denoise.py: denoise_distilled returns f32 latents (#76)
- engines/vlm.py: SSRF http(s) image fetch dropped (#65), special
  token leak fix (#68), cancelable timeout (#71)
- routes/anthropic.py: extract_multimodal_content call signature (#63)
- scheduler/ngram_spec.py: rollback trim logic (#64)
- scheduler/spec_decode.py: trim_prompt_cache hybrid no-op (#66)
- video/ltx2/video_vae/decoder.py: remove hardcoded 128 (#67)
- utils/image.py: image fetch hardening

Verification: 631 directly-related tests pass, zero regressions
(stash comparison: completions test files 70->42 failures, net -28).
CHANGELOG updated.

+++ fusion_mlx/video/ltx2/denoise.py
@@ -144,7 +144,12 @@ def denoise_distilled(
         if verbose:
             logger.info("step %d/%d", i + 1, num_steps)
 
-    return latents.astype(dtype), audio_latents.astype(dtype) if enable_audio else None
+    # Return in float32, not the input bf16 dtype. The loop promotes latents
+    # to f32 (line 33) for numerical stability and all denoise math runs in
+    # f32; downcasting the final latent back to bf16 would discard the
+    # precision distilled models (few steps) depend on. The VAE decoder
+    # consumes f32. (code-review #76)
+    return latents.astype(mx.float32), audio_latents.astype(mx.float32) if enable_audio else None
 
 
 def denoise_dev(
